test: Update test cases
