+++ api.planx.uk/inviteToPay/sendPaymentEmail.test.ts
@@ -90,10 +90,8 @@ describe("Send email endpoint for invite to pay templates", () => {
             .send(missingPaymentRequestId)
             .expect(400)
             .then((response) => {
-              expect(response.body).toHaveProperty(
-                "error",
-                `Failed to send "${template}" email. Required \`paymentRequestId\` missing`,
-              );
+              expect(response.body).toHaveProperty("issues");
+              expect(response.body).toHaveProperty("name", "ZodError");
             });
         });
